requirements.txt
djangorestframework==3.9.4
django==2.2.2